+++ src/pages/_app.tsx
@@ -102,6 +102,23 @@ export default function MyApp({ Component, pageProps: { session, ...pageProps } 
                 <Script src="/dsfr.module.min.js" strategy="lazyOnload" type="module"></Script>
                 <Script noModule src="/dsfr.nomodule.min.js" strategy="lazyOnload" type="text/javascript"></Script>
                 {process.env.NODE_ENV !== "development" && <Script src="/smarttag.js" strategy="beforeInteractive" />}
+                {process.env.NODE_ENV !== "development" && process.env["NEXT_PUBLIC_MATOMO_URL"] && (
+                  <>
+                    <Script id="matomo-init" strategy="afterInteractive">
+                      {`
+                        var _paq = window._paq = window._paq || [];
+                        _paq.push(['trackPageView']);
+                        _paq.push(['enableLinkTracking']);
+                        (function() {
+                          var u="${process.env["NEXT_PUBLIC_MATOMO_URL"]}/";
+                          _paq.push(['setTrackerUrl', u+'matomo.php']);
+                          _paq.push(['setSiteId', '${process.env["NEXT_PUBLIC_MATOMO_SITE_ID"]}']);
+                        })();
+                      `}
+                    </Script>
+                    <Script src={`${process.env["NEXT_PUBLIC_MATOMO_URL"]}/matomo.js`} strategy="afterInteractive" />
+                  </>
+                )}
               </DependenciesProvider>
             </ComparaisonContextProvider>
           </RechecheAvanceeContextProvider>