+++ CHANGELOG.md
@@ -5,13 +5,14 @@
 ### Bug Fixes
 
 ### Features
+
 - hel-1023: [Paramétrage]Historique des messages
 - hel-1022: [Paramétrage] Mise à jour du message d'accueil
 - hel-1026: [Paramétrage]Affichage du message
-
 - hel-1033: Retrait du message d’erreur en cas de compte existant dans le formulaire de création de compte
 - hel-1044: Ajouter la donnée etablissement principale/secondaire dans l’export des EJ
 - hel-1036: [RGPD] exercer les droits en matière de données personnelles
+- hel-1039: Ajout de Matomo
 
 ## [1.8.8](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.7...release/1.8.8) (2026-05-20)